updated loggedout homepage view
