+++ client/src/index.scss
@@ -1,19 +1,18 @@
 @import url("https://fonts.googleapis.com/css2?family=Outfit:wght@300&family=Raleway:wght@200;300;400;500;600;700&display=swap");
 
 body {
-  width: 100%;
-  height: 100%;
-  margin: 0;
-  padding: 0;
-  scroll-behavior: smooth;
+	width: 100%;
+	height: 100%;
+	margin: 0;
+	padding: 0;
+	scroll-behavior: smooth;
 }
 
 h1 {
-  margin: 0;
+	margin: 0;
 }
 
 * {
-  font-family: "Raleway";
-  max-width: "100% !important";
-  // display: "block";
+	font-family: "Raleway";
+	// display: "block";
 }
