+++ client/src/index.tsx
@@ -7,8 +7,8 @@ import "./index.scss";
 
 import Navbar from "./components/Navbar";
 import App from "./pages/App";
-import Register from "./pages/Register";
-import Login from "./pages/Login";
+// import Register from "./pages/Register";
+// import Login from "./pages/Login";
 import Profile from "./pages/Profile";
 import Items from "./pages/Items";
 import ItemsList from "./components/ItemsList";
@@ -18,33 +18,33 @@ import ViewProfile from "./components/ViewProfile";
 import EditProfile from "./components/updateProfile";
 
 export default function Index() {
-  ReactSession.setStoreType("localStorage");
-  return (
-    <BrowserRouter>
-      <Routes>
-        <Route path="/" element={<Navbar />}>
-          <Route index element={<App />} />
-          <Route path="register" element={<Register />} />
-          <Route path="login" element={<Login />} />
-          <Route path="profile" element={<Profile />}>
-            <Route index element={<ViewProfile />} />
-            <Route path="update" element={<EditProfile />} />
-            <Route path="*" element={<App />} />
-          </Route>
-          <Route path="items" element={<Items />}>
-            <Route index element={<ItemsList />} />
-            <Route path="add" element={<AddItem />} />
-            <Route path="update" element={<UpdateItem />} />
-          </Route>
+	ReactSession.setStoreType("localStorage");
+	return (
+		<BrowserRouter>
+			<Routes>
+				<Route path="/" element={<App />} />
+				<Route path="/" element={<Navbar />}>
+					{/* <Route path="register" element={<Register />} />
+					<Route path="login" element={<Login />} /> */}
+					<Route path="profile" element={<Profile />}>
+						<Route index element={<ViewProfile />} />
+						<Route path="update" element={<EditProfile />} />
+						<Route path="*" element={<App />} />
+					</Route>
+					<Route path="items" element={<Items />}>
+						<Route index element={<ItemsList />} />
+						<Route path="add" element={<AddItem />} />
+						<Route path="update" element={<UpdateItem />} />
+					</Route>
 
-          <Route path="*" element={<App />} />
-        </Route>
-      </Routes>
-    </BrowserRouter>
-  );
+					<Route path="*" element={<App />} />
+				</Route>
+			</Routes>
+		</BrowserRouter>
+	);
 }
 
 const root = ReactDOM.createRoot(
-  document.getElementById("root") as HTMLElement
+	document.getElementById("root") as HTMLElement
 );
 root.render(<Index />);
