+++ client/src/pages/app.scss
@@ -1,123 +1,99 @@
 @import "../global.scss";
 
 .home {
-  background-color: $whiteColor;
-  width: 100%;
-  height: 100vh;
-  position: fixed;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
+	background-color: rgb(136 161 144);
+	width: 100%;
+	height: 100vh;
 
-  .wrapper {
-    h1 {
-      font-weight: 800;
-      font-size: 600%;
-      margin-top: 100px;
-      margin-bottom: 50px;
-      text-align: center;
-      color: $darkGrayColor;
+	.wrapper {
+		width: 100%;
+		height: 100vh;
+		position: fixed;
+		display: flex;
+		flex-direction: row;
+		justify-content: center;
+		align-items: center;
 
-      @include mobile {
-        font-size: 300%;
-      }
-    }
+		.left {
+			width: 50%;
+			padding: 5px;
+			h1 {
+				font-weight: 800;
+				font-size: 500%;
+				margin-top: 100px;
+				margin-bottom: 50px;
+				padding: 0;
+				text-align: center;
+				color: rgb(255 255 255);
 
-    .loggedout {
-      display: none;
+				@include mobile {
+					font-size: 300%;
+				}
+			}
 
-      &.active {
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        margin-top: 10px;
-        margin-bottom: 100px;
-        text-align: center;
+			h2 {
+				font-weight: 500;
+				font-size: 300%;
+				margin-top: 0px;
+				margin-bottom: 50px;
+				text-align: center;
+				color: rgb(221 221 221);
 
-        @include mobile {
-          flex-direction: column;
-        }
+				@include mobile {
+					font-size: 300%;
+				}
+			}
+		}
 
-        a {
-          text-decoration: none;
+		.right {
+			width: 40%;
+			max-width: 500px;
+			background-color: white;
+			border-radius: 10px;
+			margin: 5px 10px;
+			padding: 5px;
 
-          h3 {
-            color: $whiteColor;
-            background-color: $grayColor;
-            border-radius: 8px;
-            font-weight: 700;
-            font-size: 250%;
-            padding: 4px 10px;
-            margin: 10px 0px;
+			.loggedout {
+				display: none;
 
-            @include mobile {
-              font-size: 180%;
-              text-align: center;
-            }
-          }
-        }
+				&.active {
+					display: flex;
+					flex-direction: column;
+					justify-content: center;
+					align-items: center;
+					margin: 0px 10px;
 
-        span {
-          width: 10px;
-          height: 50px;
-          margin: 0px 30px;
+					@include mobile {
+						flex-direction: column;
+					}
 
-          @include mobile {
-            width: 80%;
-            height: 4px;
-            margin: 6px 0px;
-          }
-        }
-      }
-    }
+					.login-actions {
+						width: 55%;
+						padding: 10px 5px;
+						margin-bottom: 20px;
+						display: flex;
+						flex-direction: row;
+						justify-content: space-around;
+						a {
+							text-decoration: none;
+							cursor: pointer;
+							.action {
+								color: black;
+								font-weight: 700;
+								font-size: 200%;
 
-    .loggedin {
-      display: none;
+								&.active {
+									border-bottom: 3px solid black;
+								}
+							}
+						}
+					}
 
-      &.active {
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        margin-top: 10px;
-        margin-bottom: 100px;
-        text-align: center;
-
-        @include mobile {
-          flex-direction: column;
-        }
-
-        a {
-          text-decoration: none;
-
-          h3 {
-            color: $whiteColor;
-            background-color: $grayColor;
-            border-radius: 8px;
-            font-weight: 700;
-            font-size: 250%;
-            padding: 4px 10px;
-            margin: 10px 0px;
-
-            @include mobile {
-              font-size: 180%;
-              text-align: center;
-            }
-          }
-        }
-
-        span {
-          width: 10px;
-          height: 50px;
-          margin: 0px 30px;
-
-          @include mobile {
-            width: 80%;
-            height: 4px;
-            margin: 6px 0px;
-          }
-        }
-      }
-    }
-  }
+					.forms {
+						display: block;
+					}
+				}
+			}
+		}
+	}
 }
