+++ client/src/pages/register.scss
@@ -1,135 +1,135 @@
 @import "../global.scss";
 
-.register {
-  background-color: $whiteColor;
-  width: 100%;
-  height: 100vh;
-  position: fixed;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
+.register1 {
+	background-color: $whiteColor;
+	width: 100%;
+	height: 100vh;
+	position: fixed;
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	align-items: center;
 
-  form {
-    .form {
-      background-color: $beigeColor;
-      padding: 40px 80px;
-      width: 40%;
+	form {
+		.form {
+			background-color: $beigeColor;
+			padding: 40px 80px;
+			width: 40%;
 
-      h1 {
-        font-weight: 700;
-        font-size: 300%;
-        text-align: center;
-      }
+			h1 {
+				font-weight: 700;
+				font-size: 300%;
+				text-align: center;
+			}
 
-      .formIssue {
-        display: none;
+			.formIssue {
+				display: none;
 
-        &.active {
-          display: block;
-          font-weight: 200;
-          font-size: 100%;
-          color: $errorColor;
-        }
-      }
+				&.active {
+					display: block;
+					font-weight: 200;
+					font-size: 100%;
+					color: $errorColor;
+				}
+			}
 
-      .fields {
-        padding-top: 40px;
-        padding-bottom: 20px;
+			.fields {
+				padding-top: 40px;
+				padding-bottom: 20px;
 
-        .field {
-          width: 98%;
+				.field {
+					width: 98%;
 
-          label {
-            font-weight: 300;
-            font-size: 130%;
-            padding-top: 30px;
-          }
+					label {
+						font-weight: 300;
+						font-size: 130%;
+						padding-top: 30px;
+					}
 
-          .issue {
-            display: none;
+					.issue {
+						display: none;
 
-            &.active {
-              display: block;
-              font-weight: 200;
-              font-size: 90%;
-              color: $errorColor;
-              padding-top: 4px;
-              padding-bottom: 3px;
-              padding-left: 5px;
-              margin: 0px;
-              position: inherit;
-            }
-          }
+						&.active {
+							display: block;
+							font-weight: 200;
+							font-size: 90%;
+							color: $errorColor;
+							padding-top: 4px;
+							padding-bottom: 3px;
+							padding-left: 5px;
+							margin: 0px;
+							position: inherit;
+						}
+					}
 
-          input {
-            width: 98%;
-            height: 30px;
-            font-size: 120%;
-            margin: 8px 3px 15px 3px;
-          }
-        }
+					input {
+						width: 98%;
+						height: 30px;
+						font-size: 120%;
+						margin: 8px 3px 15px 3px;
+					}
+				}
 
-        .button {
-          width: 100%;
-          display: flex;
-          justify-content: center;
-          flex-direction: column;
-          align-items: center;
-          margin-top: 15px;
+				.button {
+					width: 100%;
+					display: flex;
+					justify-content: center;
+					flex-direction: column;
+					align-items: center;
+					margin-top: 15px;
 
-          button {
-            background: none;
-            color: inherit;
-            border: none;
-            padding: 0;
-            font: inherit;
-            cursor: pointer;
-            outline: inherit;
+					button {
+						background: none;
+						color: inherit;
+						border: none;
+						padding: 0;
+						font: inherit;
+						cursor: pointer;
+						outline: inherit;
 
-            span::before {
-              content: "";
-              position: absolute;
-              top: 0;
-              left: 0;
-              bottom: 0;
-              right: 0;
-              z-index: -1;
-              background-color: $oliveColor;
-              color: $whiteColor;
+						span::before {
+							content: "";
+							position: absolute;
+							top: 0;
+							left: 0;
+							bottom: 0;
+							right: 0;
+							z-index: -1;
+							background-color: $oliveColor;
+							color: $whiteColor;
 
-              transition: transform 300ms ease-in-out;
-              transform: scaleX(0);
-              transform-origin: left;
-            }
+							transition: transform 300ms ease-in-out;
+							transform: scaleX(0);
+							transform-origin: left;
+						}
 
-            span:hover::before,
-            span:focus::before {
-              transform: scaleX(1);
-            }
+						span:hover::before,
+						span:focus::before {
+							transform: scaleX(1);
+						}
 
-            span {
-              display: block;
-              padding: 5px 10px;
-              font-weight: 700;
-              font-size: 14px;
-              box-sizing: border-box;
-              color: $oliveColor;
-              position: relative;
-              border: 2px solid $primaryColor;
-              z-index: 1;
-              transition: color 300ms ease-in-out;
-              cursor: pointer;
-            }
+						span {
+							display: block;
+							padding: 5px 10px;
+							font-weight: 700;
+							font-size: 14px;
+							box-sizing: border-box;
+							color: $oliveColor;
+							position: relative;
+							border: 2px solid $primaryColor;
+							z-index: 1;
+							transition: color 300ms ease-in-out;
+							cursor: pointer;
+						}
 
-            span:hover,
-            span:focus {
-              color: $whiteColor;
-              border: 2px solid $oliveColor;
-            }
-          }
-        }
-      }
-    }
-  }
+						span:hover,
+						span:focus {
+							color: $whiteColor;
+							border: 2px solid $oliveColor;
+						}
+					}
+				}
+			}
+		}
+	}
 }
